Stubbed headers and implementation for pyKvl classes
