+++ GEMS2/CMakeLists.txt
@@ -1,4 +1,4 @@
-CMAKE_MINIMUM_REQUIRED ( VERSION 3.5 )
+CMAKE_MINIMUM_REQUIRED(VERSION 3.5)
 
 PROJECT(GEMS)
 
@@ -6,52 +6,58 @@ INCLUDE_DIRECTORIES(${GEMS_SOURCE_DIR})
 
 # Find ITK
 FIND_PACKAGE(ITK)
-IF(ITK_FOUND)
-   INCLUDE(${ITK_USE_FILE})
-ELSE(ITK_FOUND)
-   MESSAGE(FATAL_ERROR
-           "Cannot build without ITK.  Please set ITK_DIR.")
-ENDIF(ITK_FOUND)
+IF (ITK_FOUND)
+    INCLUDE(${ITK_USE_FILE})
+ELSE (ITK_FOUND)
+    MESSAGE(FATAL_ERROR
+            "Cannot build without ITK.  Please set ITK_DIR.")
+ENDIF (ITK_FOUND)
 
 # Enable C++11 syntax
-set (CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 11)
 
 
 add_subdirectory(pybind11)
-pybind11_add_module( GEMS2Python
-             GEMS2Python.cxx
-             kvlAtlasMeshCollection.cxx
-             gzstream.C
-             kvlAtlasMeshCollectionValidator.cxx
-             kvlAtlasParameterEstimator.cxx
-             kvlAtlasMeshSmoother.cxx
-             kvlCompressionLookupTable.cxx
-             kvlCroppedImageReader.cxx
-             #kvlAtlasMeshAveragingConjugateGradientOptimizer.cxx
-             kvlAtlasMeshRasterizor.cxx
-             kvlAtlasMeshAlphaDrawer.cxx
-             kvlAtlasMeshMultiAlphaDrawer.cxx
-             kvlAtlasMeshVisitCounter.cxx
-             kvlAtlasMeshStatisticsCollector.cxx
-             kvlAtlasMeshLabelImageStatisticsCollector.cxx
-             kvlAtlasMeshProbabilityImageStatisticsCollector.cxx
-             kvlAtlasMeshDeformationOptimizer.cxx
-             kvlAtlasMeshDeformationFixedStepGradientDescentOptimizer.cxx
-             kvlAtlasMeshDeformationGradientDescentOptimizer.cxx
-             kvlAtlasMeshDeformationConjugateGradientOptimizer.cxx
-             kvlAtlasMeshDeformationLBFGSOptimizer.cxx
-             kvlAtlasMeshPositionCostAndGradientCalculator.cxx
-             kvlAtlasMeshToLabelImageCostAndGradientCalculator.cxx
-             kvlAtlasMeshToIntensityImageCostAndGradientCalculator.cxx
-             kvlConditionalGaussianEntropyCostAndGradientCalculator.cxx
-             kvlMutualInformationCostAndGradientCalculator.cxx
-             kvlHistogrammer.cxx
-             kvlAtlasMeshSummaryDrawer.cxx
-             itkMGHImageIO.cxx
-             itkMGHImageIOFactory.cxx )
+pybind11_add_module(GEMS2Python
+        GEMS2Python.cxx
+        pyKvlCalculator.cxx
+        pyKvlImage.cxx
+        pyKvlMesh.cxx
+        pyKvlTransform.cxx
+        pyKvlOptimizer.cxx
+        kvlAtlasMeshCollection.cxx
+        gzstream.C
+        kvlAtlasMeshCollectionValidator.cxx
+        kvlAtlasParameterEstimator.cxx
+        kvlAtlasMeshSmoother.cxx
+        kvlCompressionLookupTable.cxx
+        kvlCroppedImageReader.cxx
+        #kvlAtlasMeshAveragingConjugateGradientOptimizer.cxx
+        kvlAtlasMeshRasterizor.cxx
+        kvlAtlasMeshAlphaDrawer.cxx
+        kvlAtlasMeshMultiAlphaDrawer.cxx
+        kvlAtlasMeshVisitCounter.cxx
+        kvlAtlasMeshStatisticsCollector.cxx
+        kvlAtlasMeshLabelImageStatisticsCollector.cxx
+        kvlAtlasMeshProbabilityImageStatisticsCollector.cxx
+        kvlAtlasMeshDeformationOptimizer.cxx
+        kvlAtlasMeshDeformationFixedStepGradientDescentOptimizer.cxx
+        kvlAtlasMeshDeformationGradientDescentOptimizer.cxx
+        kvlAtlasMeshDeformationConjugateGradientOptimizer.cxx
+        kvlAtlasMeshDeformationLBFGSOptimizer.cxx
+        kvlAtlasMeshPositionCostAndGradientCalculator.cxx
+        kvlAtlasMeshToLabelImageCostAndGradientCalculator.cxx
+        kvlAtlasMeshToIntensityImageCostAndGradientCalculator.cxx
+        kvlConditionalGaussianEntropyCostAndGradientCalculator.cxx
+        kvlMutualInformationCostAndGradientCalculator.cxx
+        kvlHistogrammer.cxx
+        kvlAtlasMeshSummaryDrawer.cxx
+        itkMGHImageIO.cxx
+        itkMGHImageIOFactory.cxx
+        )
 
-TARGET_LINK_LIBRARIES( GEMS2Python
-                      PRIVATE ${ITK_LIBRARIES}
-                      PRIVATE ${ZLIB_LIBRARIES}
-                     )
+TARGET_LINK_LIBRARIES(GEMS2Python
+        PRIVATE ${ITK_LIBRARIES}
+        PRIVATE ${ZLIB_LIBRARIES}
+        )
