+++ GEMS2/pyKvlTransform.h
@@ -1,0 +1,9 @@
+#ifndef GEMS_PYKVLTRANSFORM_H
+#define GEMS_PYKVLTRANSFORM_H
+
+#include "kvlCroppedImageReader.h"
+
+typedef kvl::CroppedImageReader::TransformType  TransformType;
+typedef TransformType::Pointer TransformPointer;
+
+#endif //GEMS_PYKVLTRANSFORM_H
